Changed the map colors
This shit was harder than I thought, google maps has a half assed
HSL implementation which requires black magic or something. After
hours of splitting my brains I found out about this SO question:
<http://stackoverflow.com/questions/3472040/rgb-value-to-hsl-converter>
whose piece-of-shit first answer was a load of bullcrap; but the second answer:
<http://googlemapscolorizr.stadtwerk.org/>
worked flawlessly. Too bad I noticed it after an hour of calculating shit
in gimp from the first answer trying to make heads or tails of it.

FUCK THAT SHIT.